fixed missing compilation issues

+++ automation/bearssl-0.6/v2/harness/encryption_util.c
@@ -1,6 +1,6 @@
 #include "encryption_util.h"
 
-void fill_words_deterministically(uint32_t* const dest, size_t n, uint8_t seed) {
+void fill_words_deterministically(uint32_t* const dest, size_t n, uint32_t seed) {
     uint32_t v = seed;
     for (size_t i = 0; i < n; i++) {
         v = v * 1664525u + 1013904223u;  // simple LCG
